requirements.txt
mysql-connector-python
Flask
flask-cors